decode_doa6_soft: Dump bone -> vertex group maps in standalone
This is to try to work out how the bones are mapped to different vertex
groups in each region of the mesh. The hope is that we might be able to
turn this into a .vgmap file that can be used by the Blender addon
script to give us text bone names rather than using the vertex group
indices, and possibly even alter the mapping in the g1m files to add
any missing bones to sections of mesh that need them, but for now this
is just exploratory.