decode_doa6_soft: Improve vertex placement outside of the grid
This relaxes the restriction that vertices must be within the soft body
grid to calculate usable soft body weights and position them accurately.

If a vertex is outside of the grid and four surrounding nodes have been
located the script will now search for an additional four nodes near
those forming a cube and simply allow the interpolation to find a result
that falls outside of that cube.