blender addons: Export DOA6 bone to VG map & support text VGs
The decode_doa6_soft script will now parse the .oid file to get the
mesh component names, the G1MS section to map bones IDs to those names,
and the G1MG section to map vertex groups to bone IDs and their names.

It will then write out a .vgmap file for each of the mesh regions in the
g1m file mapping bone names to vertex groups that can be imported into
Blender with the blender_3dmigoto addon.

The Blender 3Dmigoto addon is extended to support vgmap files with text
keys, and an option added to rename existing vertex groups on import to
match.

The final result is that we can now use the text names for the vertex
groups of DOA6 meshes, and still export them with the correct vertex
groups.