Blender 3DMigoto Addon: Import & Export vgmap files with raw buffers
If a vertex group mapping is in use this will export it as a .vgmap
file next to the raw .vb, .ib and .fmt files. When importing raw buffers
this will check for any .vgmap files next to them and import it.

This allows DOA6 meshes extracted from the game files to be imported and
exported with vertex group names with minimal effort. Currently the
vgmap file must be extracted by running decode_doa6_soft.py over a .g1m
file, but the idea is that a future version of vagonumero13's g1m_export
utility could also extract these.