Blender Addon: Add operator to update VGMap with new vertex groups
This adds an operator to update a vertex group mapping attached to an
object assigning new vertex groups mappings to any new named
(non-numeric) vertex groups.

This is intended to be used in DOA6 to transfer bones from one section
of mesh to another, e.g. where the body mesh is missing foot bones due
to being covered by boots.