Blender Addon: Clean up unreferenced vertex groups
This adds an option when importing a vgmap file to remove any vertex
groups that are not listed in the file. This is intended to make DOA6
meshes a little easier to work with, since they have a vertex group step
of 3, so 2 out of every 3 vertex groups are blank.

Naturally this will require a complete vgmap file, not a partial one.