Blender Addon: Fix .vgmaps applying after failed .vb import
If .vgmap files are present when importing raw .vb/ib/fmt files the
script would try to apply them to the active object after import without
checking if the import had succeeded, either throwing an error or
potentially assigning them to the wrong object.

This returns the imported objects from import_3dmigoto and vgmap files
will only be applied if the import was successful.