DOA6 Blender Addon: Add --import-vgmap option
This adds an option to import the vgmap files back into the g1m mesh. It
simply replaces all existing mappings with the new ones without
bothering with de-duplication (sets all bone map IDs in the surface
section as 0 through the number of surfaces).

The G1MF section also had to be modified, as it contained the number of
maps and total number of mapped bones that needs to be updated to avoid
a crash.

In layman's terms: This allows bones from one mesh section to be used in
another that lacked them, e.g. transferring foot bones from a boot mesh
to a body mesh.