DOAXVV: Update hashes for 2019-04-24 update
A number of skin & clothing shaders have been updated to use the new
parameter that was added last December, and as such have had their
hashes change. Add the new hashes to restore costume mods to working
order.

Leave the old hashes in as well, since the Steam version has not yet
received this update.