shader database: Add mirroring (not indexing) of 7z archives as well
The old DX9 shader database is still up and running, and despite never
having been updated for DX11 fixes does still mirror them which is
useful in its own right. However, it only downloads the fixes if they
are either in zip or rar archives, but 7z is now a fairly common choice
as well and would be beneficial to mirror 7z fixes as well.

This simply adds 7z to the list of extensions that will be mirrored, but
makes no attempt to index the contained shaders.