Shader Database: Add support for Dropbox URLs
Dropbox URLs direct to a useless web interface by default, but editing
the URL to specify dl=1 allows us to bypass that and go to the thing we
actually requested. This adds a function to alter these URLs as
necessary.

Unfortunately Dropbox neglects to give as a Last-Modified header, so we
have no way to know if the file has been updated or not. It's possible
this might be implied by the rest of the URL, but I'm not familiar
enough with it to say. Standards violating websites like these are evil.