Team Sonic Racing: Fix Shadows
This takes a view-space correction to line up the shadows, but the
shader in question lacks the projection matrix to do this. This derives
the projection matrix by inverting the world-view matrix and multiplying
it by the world-view-projection matrix. This is currently done via a
separate compute shader, but the Unity 4 pattern could be included in
the shadow shader itself for better optimisation.