Team Sonic Racing: Fix game bug - shader using white depth texture
In the second track at one point there is a line of light crossing
through a shadow at an impossible angle (does not match direction of the
sun), and attempting to stereo correct it results in it moving to
infinity... On closer inspection, the depth texture the shader is
reading from appears to be solid white, which suggests to me that it
might be a game bug.  Assign the depth target to this depth texture,
which subtly changes the lighting but more importantly - removes the
impossible line of light.