Team Sonic Racing: Force full screen + add HUD toggle
TSR seems a bit iffy as to whether it engages exclusive or borderless
full screen - it seems to start in borderless, but alt+tab on a monitor
seems to switch it to exclusive, yet over HDMI it seems we need to force
exclusive mode.

Also add a debug HUD toggle key to make it a little easier to focus on
the graphics.