Team Sonic Racing: Fix Mirror Mode tracks
Tracks in this game can be played in regular or mirror mode, where the
mirroring swaps left and right and is done as a post-processing effect
by flipping the texture coordinates passed to an existing
post-processing shader. Mirroring the world in this way will also mirror
the stereo correction, effectively inverting the eyes.

To detect this, we therefore have to compare the texture coordinates
used in that shader with the screen position and note if they are
reversed, then use a stereo2mono to swap the eyes around. Since we can't
currently check the contents of vertex buffers from the CPU this test
needs to be done on the GPU side and the stereo2mono will have to be
unconditional with a custom shader deciding whether to apply the eye
swap or not.

There is an additional problem however - the post-processing shader used
for mirroring does not simply draw a full screen quad as two triangles,
but rather uses 16 triangles zig-zagged across the screen. The four
center-most triangles will include vertices right at the dead-center of
the screen, where the mirrored and non-mirrored texture coordinate will
be 0.5, so these vertices cannot determine if mirroring is in effect or
not.

To get around this issue we need to use a different set of vertices that
don't include the center of the screen (alternatively we could copy the
vertex buffer to a texture slot, but there is a performance penalty
involved in that we would rather avoid) by passing in a modified index
buffer. Here we merge the four center triangles down into two larger
triangles so that no vertex lies dead center, though there certainly are
other options that could achieve the same.