Team Sonic Racing: Optimise shadow fix
The shadow shader has the world-view and world-view-projection matrices
and needs the horizontal FOV component in the projection matrix, which
is essentially the same challenge as Unity titles. In Unity, we used
optimised assembly to derive this component, so we can do the same here
saving the cost of running a compute shader and passing the matrices
around.

The motivation is trying to minimise our performance hit on this game,
which already seems to struggle somewhat.

Also fixes a long standing typo in the version of this code in one of
the old Unity 4 template shaders that was using an uninitialised
temporary register instead of the correct constant register.