Team Sonic Racing: Another shot at moving specular reflections to infinity
Seems like the infinity adjustment still wasn't quite right, as evident
in some parts of some tracks that the reflection could still rest above
the surface. Examining the vertex shader revealed a multiplication by a
value from the constant buffer, which when added to the adjustment seems
to indeed move the reflections to true infinity, verified with a very
low convergence test.