Team Sonic Racing: Fix Ambient Occlusion shadow artefacts
This applies a stereo correction to the AO shader to fix artefacts in
shadow. As a side note, this is the shader previously identified as
being implicated in the false shadow game bug, though that is separate
from this, and in fact that bug would have masked this issue since the
depth buffer was blank giving AO nothing to do before.