Team Sonic Racing: Move specular eye position variant to ShaderRegex
The glass surface shader a little further in the cloud stage exhibited
another instance of the specular eye position variant, but this time it
was actually the surface shader rather than a purely post processing
shader. This is reminiscent of specular highlights embedded in material
shaders in other games such as Unity 5, and typically ends up being too
many to fix by hand. Convert the fix into a ShaderRegex to hopefully
catch all instances.