Team Sonic Racing: Correct matrix-multiply method for specular highlights
Some of the specular highlights were correcting the wrong way in
different parts of the same track, suggesting there was something still
wrong with the specular correction. The method came from Unity, which
also uses "row_major" matrices, but they are transposed compared to this
game, so the dot product matrix multiply method is needed here.

This seems to have all specular highlights adjusted in the right
direction now.