Team Sonic Racing: Add a HUD adjustment while driving
This pushes the HUD into depth while driving to make it easier to see
team mate nameplates, incoming item alerts, etc.

This game is problematic for doing sophisticated HUD adjustments for a
number of reasons - the split screen and mirror mode make using a depth
buffer adjustment tricky, since the most recent depth buffer may not be
the one we need, and we would need to take into account mirror mode to
flip the sample coordinates. The game also bunches numerous textures
together and updates them live, making it impossible to identify
specific textures from a command list or even to reliably detect any
texture that might indicate we are racing... The final kick in the teeth
is the mouse cursor that we would almost be able to use to determine
when we are in a race... except half the time the game neglects to hide
it at all.

So, this ends up going for a different approach. The HUD adjustment is
triggered by the accelerator key, since we can reasonably expect the
player to hold that down most of the time they are racing, and we have a
short release delay to allow for cornering, with transitions to avoid
abrupt changes.

The actual adjustment we use takes advantage of the fact that the race
occurs on a road, so HUD elements near the bottom of the screen should
be near the player car, and those higher in the screen should be further
into depth. It interpolates between two depth values configurable in the
d3dx.ini based on the screen Y value (which fortunately is relative to
each individual split screen). The result is not perfect, but quite
usable.