Team Sonic Racing: Fix wrong depth in AO shader
Didn't quite notice that the matrix I was using to get the depth was an
inverse view-projection matrix, i.e. it ends up in world-space, not
view-space. Therefore, the depth value I obtained from it was garbage...
Surprisingly it still worked just fine in most directions, but created
weird lines in very specific locations.

Add the right foo to normalise the value in world space, convert back to
clip space and THEN we have the depth.