Escape from Tarkov: Fix Decals
These follow a very similar pattern to the lights in Unity, needing the
same sort of halo + ray correction in the vertex shader and a second
view-space correction in the pixel shader. Unlike lighting we can
probably rely on this always having valid matrices, so can avoid copying
data from elsewhere.

The particular halo pattern here does not get identified by hlsltool
automatically due to an unusual swizzle (output.xyzw = temp.xyww), but
since the rest of the shader needs a fix and so far there do not appear
to be additional variants on this shader it is not yet worth extending
that just yet.