Tarkov: Partial SSR fix
This forces the SSR render target to stereo and applies corrections
around the projection & inverse projection matrices, which moves the SSR
reflection to surface depth.

From this point it is usually a matter of adjusting the camera position
to then place the reflections at accurate depth, however this shader
does not use the camera position. It might be doing view-space
calculations and assuming the camera is at 0,0,0, but that will take
more effort to find and correct - I might come back to this later once
other more serious issues are fixed.

There also appears to be a separate issue with the reflection brightness
between the eyes, possibly from another shader.