Blender 3DMigoto addon: Initial support for Blender 2.80
This adds support for Blender 2.80, which has made numerous API changes
with little regards to backwards compatibility, as is typical of many
Python projects (to be fair, the language doesn't do anything to
encourage developers to maintain a stable API / discourage developers
from breaking a public API, and in some cases they have actually made
provisions for a deprecation phase... but since this isn't done for all
API changes it doesn't actually help all that much).

This attempts to retain compatibility with both Blender 2.79 and 2.80,
using various checks and wrappers to select the appropriate API to use.

orientation_helper_factory is replaced with a new decorator. The script
is updated to the new API, however the old API is also retained, with
a dummy class/decorator being provided for whichever API is not in use.

Import and export menus are renamed in the API. These are fetched early
and stored in our own variables.

Various wrappers have been incorporated from the below site, to
automatically transform properties into annotations and abstract certain
APIs:
https://theduckcow.com/2019/update-addons-both-blender-28-and-27-support/

This is enough to get frame analysis and raw buffer imports working in
Blender 2.80. A few additional changes will likely be required for other
features.

XXX CHECKME: show_edge_sharp moved in 2.80, but I can't actually recall
what it does and have a feeling it was unimportant, so have simply
disabled it. Need to double check this.