Unity template: Force stereo depth buffer & recommend stereo compute shaders
One of the depth buffers in Green Hell does not get automatically
stereoised by the driver, and I've seen this happen in several other
Unity games recently, tricking some shaderhackers into thinking that
halo problems had not been fixed when they had. Add a fuzzy
TextureOverride to the template to hopefully match all of these cases
and force the affected depth buffers to stereo.

Green Hell also uses compute shaders, which is a first for a Unity games
as far as I'm aware. Recommend enabling stereo compute shaders in the
d3dx-excerpts file to support these. Note that this option must be
enabled in the top level d3dx.ini - it cannot be added to the unity.ini
since the profile helper does not parse included files.