Unity Template: Add a render target size check when grabbing Unity resources
Green Hell updates reflection probes roughly once a second, and when
this happens the template grabs bad matrices and depth buffers. This
adds a check to make sure the render target is rectangular when grabbing
any Unity resources to try to ensure that only the viewport matrices +
depth buffer are obtained.