DOA6: Add new main menu texture hash
Update has broken a few minor things - this fixes the main menu, but
I'll need to come back to the others when I have some time. Also noticed
the volumetric fog in A.P.O. is broken, and some lighting levels are
slightly different in each eye in some stages.