Generic Shader Extractor: Add support for DX9 shaders
This adds support for scanning unknown binary files for DX9 shaders and
extracting any that are found. They are just extracted as binary for the
moment, and will need to be run through fxc to disassemble:

    fxc.exe /dumpbin shader.bin /Fc shader.txt

DX9 shaders lack any magic bytes to identify the shader as a whole, so
this relies on locating their constant table by the CTAB magic bytes and
assuming that the shader starts 8 bytes earlier. This will work for most
DX9 shaders, though any shaders that have debug info left in or
otherwise have another section before CTAB or are missing CTAB will be
able to defeat it. DX9 shaders also lack a field indicating their size,
so this scans through the instructions to find the end.

It writes the shader to Helix Mod like directories (albeit, as binary)
with the crc32 hash used as a filename, though note that the hash has
not been verified in a real world example and could be incorrect if e.g.
the parser does not stop on the correct end of the shader.