Blender Addon: Add placeholder to remap poorly named semantics
If the game lies about its semantics the user might need to tell us how
to treat each one. This adds the beginnings of a mechanism to do this
via a map in the import routine. This does not yet handle mapping these
back on export, nor is it hooked up in the GUI.

An alternative approach that already works is to edit the semantic in
the vb*.txt / *.fmt files.