FC5: Fix ground
This game seems to render ground significantly differently to previous
FC games. The 3D Geometry without texture is rendered in one pass, then
this shader is used to render out three G buffers that will be consumed
in later lighting passes. That means that even though this isn't what we
would usually consider a post-processing shader, it has all the
hallmarks of one as it doesn't do any forwards coordinate
transformations, but does several inverse transformations in the pixel
shader, and therefore needs fixes similar to what we would ordinarily do
for shadows.