DOAXVV/geo-11: Force reflections to stereo
This matches the reflection RT/texture and forces them to stereo. For
now putting this in the d3dxdm.ini to make it geo-11 specific, as last
time I ran this game in 3D Vision Automatic the reflections were already
stereo - will need to double check if that is still the case though in
case they were broken by a game update.