DOAXVV: Tweak specular VS adjustment
This adjustment is messing up sand at sunset when the camera is at right
angles to the sun, and besides that the adjustment looks too strong
anyway. It's likely that the depth value here is wrong or otherwise that
this is not the correct fix for these reflections. The r0/o8 adjustment
seems to be the worst part of this, so disable adjusting that for now to
fix the sand.

The remaining two adjustments are much more subtle, don't mess up the
sand, and look about right on shiny materials they do affect, so leave
them in. However they don't appear to have any noticeable affect on some
other materials using this same shader, so I'm still searching for the
ultimate specular fix for this game.