DOAXVV: Another attempt at specular highlights. Looks good.
From outputting these in the pixel shader I suspect these are view-space
normals, but since the camera/eye position doesn't seem to be used in
either vertex or pixel shader (nor does the pixel shader have any depth
information or depth buffer etc) I can't apply the usual fix for these.

Experimenting with a few possible patterns I've settled on applying:

separation * convergence, because that's how much we typically adjust
the eye/camera position to fix specular highlights

* inverse projection matrix, because this is view-space

/ depth, effectively a perspective divide because otherwise this messes
up sand way too much. In this case I'm using the depth from the output
position. This may not be accurate, as we generally don't need a
perspective divide for specular fixes.

Regardless of whether this is accurate or not, it looks really good in
stereo 3D - wet skin and many shiny materials (that weren't covered by
the other specular fix) are affected.

I've disabled the adjustments to the other two view-space outputs - they
are too subtle to estimate if they are right or not, and given that the
pattern I had used with them did not work with the r0/o8 output I'd
guess they are probably wrong.

Doing this as a regex because I think with the number of altered
d3dx.ini files floating around for this game I don't want a manually
patched shader to depend on a [ShaderOverride] existing. Either a user
has this pattern in their d3dx.ini and gets the full fix or not at all.