Blender addon: Support importing split vertex buffers from binary
This should be the final piece in the puzzle to support split vertex
buffers - this can allow the script to import .buf files from frame
analysis dumps (via the buf checkbox in the .txt import dialog) and .vb*
binary files previously exported by the script.

Only lightly tested. I would judge the likelihood of this (and the
preceding two commits) causing a regression as moderate so it should go
through a few more paces before being merged to the master branch.