DOAXVV/help text shader: Add effective_dpi scaling (requires future 3DMigoto)
This adds support for using the effective DPI scaling factor to scale
the help text up on ultra high resolution displays (e.g. 4K).

This support will be added in a future version of 3DMigoto & geo-11, so
for now this is only added to the DOAXVV copy of the shader, and omitted
from the generic copy. It does have a rudimentary fallback to keep
working on current versions, but will emit warnings about effective_dpi.

Note that effective DPI support was only added in Windows 8.1, and
improved in Windows 10 1607, and may be affected by the application DPI
awareness and possible windows compatibility settings... So how well
this works in practice may vary from case to case - wider testing would
be appreciated.