Help text shader: Work around missing text when used in geo-11
For some reason the last 13 characters out of each 64 character chunk
are missing when this shader is used with geo-11. The reason is not yet
clear (almost like the geometry shader output stream is smaller than it
should be), but for now work around it by reducing the number of
characters in each chunk to 51.