CB Live view debug shader: Add effective DPI scaling
Adds the new/upcoming effective_dpi support to this shader to scale the
text up on ultra high resolution displays to visually match what it
would have displayed as on 1080p. Requires a future version of 3DMigoto
/ geo-11 to support this, otherwise it will not scale (and will warn
about effective_dpi).

Updates the debug_cb shader included with DOAXVV to latest - I wouldn't
update the help shader namespace because 3rd party mods depend on being
able to reference it's resources as is, but the debug shader should be
ok since it isn't intended for production. This also corrects the use of
the broken StereoParams resolution (that is absent in geo-11) to use the
render target resolution.