DOAXVV: Fix specular in compute shader
The vertex shader ended up being a red herring. The correct fix for
these specular highlights is in the lighting compute shader. The r2
register in the shader is used in many code paths, some of which need
the specular adjustment, others are used for shadows and are broken by
the specular adjustment. Replace all uses except those that go through a
view to shadow matrix.