DOAXVV: Fix tile lighting on point lights
This issue is rare and minor, but is evident in the fireworks photo area
- 2 point lights on the barge and one on a torch near a random tree.

This is a relatively easy fix in the lighting compute shader - the first
loop that iterates over gPointLightList is calculating tiles (it is
evident because that loop only reads the light positions, not colours,
and it writes to a thread group shared memory g6 register that is read
by later loops), while the later loops on that list are actually doing
the lighting. Apply a view-space correction to the light position in the
first loop so the tiles will surround the stereo position of the light
instead of the mono position.

From what I can tell the spot lights do not appear to use tiling (e.g.
all loops over that list read the color) and so don't need a similar
correction.