DOAXVV: Fix double-correction of Chandelier reflection in Studio
The reflection of the Chandelier light on the floor of the Studio was
too deep - seems it was affected by both the point light adjustment and
the camera position adjustment. Playing around a little it seems that
some uses of r2 should remain as the shadow/surface depth adjustment so
that this light will only receive the point light adjustment.