DOAXVV: Re-fix bathroom mirror reflection
Removing the double correction on the Chandelier light reflection also
reverted the bathroom mirror reflection to surface depth :(

So make two copies of the r6 register, one with the camera depth
correction to fix the mirror, the other with the surface depth
correction do avoid double-fixing the Chandelier reflection.

Also apply the pattern to the other two likely locations in the shader,
though I haven't seen exactly what uses those.