Blender addon: Don't treat 4D positions/normals as fatal
There are some games that store some data (not just all zeroes / ones)
in the 4th dimension of a POSITION or NORMAL semantic, e.g. these show
up in skirts in DOA6 and have also been reported in WARRIORS OROCHI 4.

Obviously positions and normals are 3D in Blender so we can't simply
import this 4th dimension, certainly not without knowing what it is, so
the plugin previously raised a fatal exception if it encountered this.

There may however be some cases where it might be useful to attempt to
import, modify and export the mesh anyway, even though we don't know
what this 4th dimension is used for - in an optimistic case the game
might not actually use the 4th dimension for anything and we can just
ignore it. There may also be some types of edits where we can just
preserve the data in the 4th dimension and write it back on export
without modification and hope that it works - that probably won't work
if vertices have been added or subdivided, but might work ok if vertices
have only been repositioned or deleted.

So, this commit downgrades this fatal exception to a warning, and stores
the 4th dimension in a vertex layer - the POSITION case already had code
to do this if the assertion was commented out, this also adds equivalent
code to the NORMALs.

Closes https://github.com/DarkStarSword/3d-fixes/issues/3