Blender Addon: Support importing+exporting with no index buffers
It's perfectly valid to not have an index buffer, though somewhat rare
as that would generally mean duplicating vertices for a triangle list
topology. The script was already written with these possibilities in
mind, but they were untested and there's a couple of minor bugs and
missing export logic that prevented it working in practice.