Blender addon: Catch exception calculating tangents
It seems that the mesh.calc_tangents() call requires a UV layer, so
export will currently fail if one is missing. It's not entirely clear
how best to handle that scenario (e.g. perhaps we should try to preserve
any tangents on import), but at least try to export anyway with warnings
that tangents/normals/lighting might be bad. Also add a warning on
import if there is no TEXCOORD semantic to import as a UV layer so the
modder will know in advance of the possible issue.