Blender addon: Add experimental support for pointlist topology
Unity and certain other games do their skinning in a pre-pass where the
mesh's vertex buffers are bound as point lists with no index buffers.
This pass outputs a new vertex buffer with skinning applied that is used
in future draw calls of that mesh (which will have the actual index
buffers assigned and use the triangle list topology).

This change allows the vertex buffers passed into this skinning step to
be imported into blender, and possibly exported back to the game to
potentially allow for some basic modifications.

Please note, that loading in these meshes from point lists is probably
NOT a good workflow - too much information is missing (edges and faces)
or lost (normals + tangents) - it may ultimately be better to do a
hybrid where we use these pre-skinned vertex buffers with the index
buffers from later draw calls, but regardless it may still be useful to
be able to work with these point lists in some cases.