Blender Addon: Add support for loading point lists with index buffers
We see this in One Piece Burning Blood - see samples attached here:

https://github.com/DarkStarSword/3d-fixes/issues/19#issuecomment-1748867466

These index buffers are pretty pointless since they just list every
vertex in order and lack the advantages that index buffers provide for
triangle list topologies that are used to avoid listing duplicate
vertices, but it is technically valid to use index buffers with point
lists, so we might as well support it.

This currently just assumes that they are the trivial case that we can
pretty much ignore as though they didn't use index buffers, and the
implementation is essentially just an assertion to verify that is the
case, and to break the assumption that index buffers will be groups of
three vertices.