Blender addon: Allow lower/mixed case semantic names
The semantic names that DirectX knows about are supposed to be
uppercase, but some games use lower or mixed case for these names, so
allow these to be treated as equivalent in the script.

It's not clear if DirectX will still "officially" recognise these
semantics, or if it would treat them as arbitrary semantics, but given
that these are "completely generic and is not uniquely interpreted by
the system" anyway and just need to match the shader inputs it really
doesn't make much difference.

Among other cases, this was brought up here:
https://github.com/DarkStarSword/3d-fixes/issues/19#issuecomment-1751690372