Blender addon: Optionally use draw call range when importing buf files
Shining Resonance Refrain uses the triangle strip topology, and has
multiple strips in the index buffers. Loading the frame analysis dumps
with "Load .buf files instead" enabled winds up with broken geometry
because it tries to load all indices and strips from the index buffer,
but this doesn't really work with strips as they have to be loaded
individually.

This internally adds a use_drawcall_index_range flag in parse_ib_bin
that restricts it to load only the indices specified in the draw call
instead of the entire buffer, allowing the geometry to be loaded
correctly.

This is not (yet) exposed in the UI, and is disabled by default. It is
largely unnecessary for end users since loading the .txt files already
takes the draw call index range into account, so this option is really
more for development.