Blender Addon: Discard index buffers that weren't used in the draw call
3DMigoto will dump whatever resources were requested in frame analysis
if they were bound to the pipeline at the time of the draw call, but
this does not necessarily mean they were actually used in the draw call.

Shining Resonance Refrain leaves index buffers bound during Draw() calls
that don't use them, which confuses the script into trying to apply
indices that aren't applicable to the draw call.

This adds a check to make sure the index buffer was used signified by
the first index + count in the header, discarding it if it was not used.