Blender Addon: Parse MS -1.#IND and similar nan/inf values
Microsoft's standard libraries encodes NaN/Infinity values in a way that
Python cannot parse. This adds a parsing routine that can handles these
values, translating them to +/- nan/inf

Seen in some of the meshes in Shining Resonance Refrain

Notably in some cases these values may be encountered in an invalid
semantic (i.e. one that shares the offset of another semantic) if an
integer is interpreted as a float, for example as seen in:
    https://github.com/bo3b/3Dmigoto/issues/166