Blender Addon: Rework FA import options to use sub-panels
This disables Blender from automatically adding the import operator
options to the import dialog so that we can add our own sub-panels to
group related options, and to grey out mutually exclusive options.

Main motivation for this was to place the "Limit to draw range" in a
sub-section of "Load .buf files instead" to clearly indicate that it is
not relevant when importing .txt files for the normal workflow.

It also finally allows me to grey out the option to load / merge related
buffers when loading a .buf file (though actually now that loading .buf
files can be limited to the draw call range it might be worth changing
the mutual exclusivity to allow those).