Blender Addon: Add support for remapping semantics
This gives the script a native way to deal with games that don't follow
standard DirectX semantic names, by allowing the user to override how
the script will interpret each semantic.

This is particularly useful for Unreal Engine games that name every
semantic "ATTRIBUTE". At the moment the user must still deduce which
attributes are what, but there's enough clues that we can probably add a
guessing feature later (e.g. POSITION is usually 1st, UE seems to group
NORMAL and TANGENT in the same buffer, TEXCOORDs are 2 component
formats, also grouped in the same buffer, blend indices and weights
always come in pairs with the indices being integers, etc).

The translations are applied at import/export time, and the before/after
renaming is added to the input layout description so that exported files
can be reimported and the same remapping will be applied.

This is a fairly substantial change, so it comes with a moderate risk of
regressions.

Allows the meshes from Unreal on this ticket to be imported without
needing to modify the headers first:
https://github.com/bo3b/3Dmigoto/issues/181