Blender Addon: Silence exception drawing the semantic remap list
Not positive what's going on here, but some of the panel draw() calls
don't have context.path_resolve = None, which causes an exception in
draw_ui_list(). It seems like draw() is called several times in a row
and the first time has this issue. Not entirely sure why it's called
multiple times - maybe I haven't implemented poll() properly and it's
trying to draw multiple times / in several places...?

Whatever the case, it seems harmless, so for now just detect and return
to avoid the exception.