Blender Addon: Add semantic info to the tooltip in the remap list
To be able to guess what we should remap the semantics to, it's helpful
to have some information about them, such as their DXGI format and which
vertex buffer they are stored in.

This will record this information from the input layout when pre-filling
the remap list and display it in the tooltip while hovering over the
semantic name.

If a semantic is manually added this info won't be present, so the tool
tip will hint to look in the menu to find the pre-fill option instead.