Blender Addon: Add backwards compatibility check for generic_ui_list.py
generic_ui_list.py looks to have only been added to Blender relatively
recently, around about version 3.5. Add a backwards compatibility check
and disable the semantic remap feature if this list helper is not
available so that we don't break users that have to use an old version
for one reason or another.

We could alternatively implement the list without relying on the helper,
but the extra code required to achieve that doesn't seem worth it IMO
considering that the remap feature is a nice to have.

This issue was noted here:
https://github.com/DarkStarSword/3d-fixes/issues/19#issuecomment-1781250808