Blender Adodn: WIP exporting .ini file to streamline re-injection
This adds a function to generate an .ini file on export that is suitable
to have 3DMigoto load all the vertex and index buffers and replace the
original mesh when the game goes to render it. It still needs some more
testing and a few final touches to write out the shader and buffer
hashes, but the main thing that needs to be decided before it can be set
loose upon the world is whether the ShaderOverride and TextureOverride
sections should be written to the same file, or at all.

The resource sections all make sense to always include, since they
contain the mundane attributes of the buffers that can be a source of
transcription errors / forgetfulness if filled out by hand.

The TextureOverride sections will usually make sense to include, but
there are some cases where a modder wants some additional control over
how the object is drawn, so it may make some sense to have this in a
separate file, optional or commented out. Also, at the moment the buffer
backup + restore code is added here, but in many cases it may be
preferable to do this in a ShaderOverride (or a CommandList called from
one) before / after the checktextureoverride.

The ShaderOverride section is probably the most controversial, since if
multiple objects are being modified and we add these sections for each
object we will quickly end up with duplicate ShaderOverride sections,
triggering warnings in 3DMigoto and running excessive
checktextureoverride sections. Therefore in most cases we probably won't
want this section included in favor of it being somewhere central like
the main d3dx.ini or another common ini file. However, it can still be
nice to have the option to include this section in some cases to be able
to quickly replace a model without having to set up an entire modding
infrastructure...